[modify] favorite adapter 완성 + 각 게시글 삭제 오류 해결 + 회원 탈퇴시 storage 폴더 전체 삭제 구현중

+++ app/src/main/java/com/cookandroid/s2_archiving/ViewAdapter.kt
@@ -2,7 +2,6 @@ package com.cookandroid.s2_archiving
 
 import android.content.Context
 import android.content.Intent
-import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -10,7 +9,6 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
-import androidx.core.content.ContextCompat.startActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
@@ -20,17 +18,16 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.storage.FirebaseStorage
-import java.security.AccessController.getContext
 
 class ViewAdapter(val viewDataList: ArrayList<PostData>, val context: Context) : RecyclerView.Adapter<ViewAdapter.CustomViewHolder>() {
 
     private var mFirebaseAuth: FirebaseAuth? = FirebaseAuth.getInstance() //파이어베이스 인증
-    private var mDatabaseRef: DatabaseReference =
-        FirebaseDatabase.getInstance().getReference("Firebase")//실시간 데이터베이스
-    private lateinit var fbStorage: FirebaseStorage
+    private var mDatabaseRef: DatabaseReference = FirebaseDatabase.getInstance().getReference("Firebase")//실시간 데이터베이스
 
-    private var activity: MainActivity? = null//메인에 함수 부르기 위해 선언하기
     private var storage : FirebaseStorage? = FirebaseStorage.getInstance()
+
+    // flag
+    private var flag:Int =0
 
     inner class CustomViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val viewImage = itemView.findViewById<ImageView>(R.id.view_img)
@@ -56,6 +53,7 @@ class ViewAdapter(val viewDataList: ArrayList<PostData>, val context: Context) :
             Glide.with(holder.itemView)
                 .load(viewDataList[position].postPhotoUri)
                 .into(holder.viewImage)
+            flag = 1
         }
         holder.viewDate.text = viewDataList[position].postDate
         holder.viewSpecial.text = viewDataList[position].postDateName
@@ -84,8 +82,10 @@ class ViewAdapter(val viewDataList: ArrayList<PostData>, val context: Context) :
             val friendId = viewDataList[position].postFriendId
             val imageFileName = "IMAGE_" + postId + "_postImage_by"+friendId+".png"
 
-            storage?.reference?.child("${mFirebaseAuth?.currentUser!!.uid}")?.child(imageFileName)?.delete()?.addOnSuccessListener {
-                Log.d("storage","이미지 삭제완료")
+            if(flag==1){ // 만약 이미지를 포함해서 저장했다면
+                storage?.reference?.child("${mFirebaseAuth?.currentUser!!.uid}")?.child(imageFileName)?.delete()?.addOnSuccessListener {
+                    Log.d("storage", "이미지 삭제완료")
+                }
             }
 
             mDatabaseRef.child("UserPosts").child("${mFirebaseAuth!!.currentUser!!.uid}").child(postId).removeValue().addOnSuccessListener {
